requirements1.txt
pandas
numpy
dvc
keras
scikit-learn
matplotlib
seaborn
tensorflow
joblib
PyYaml
streamlit
h5py
keras_preprocessing
mlflow
opencv-python
